+++ src/core/BUILD.gn
@@ -499,6 +499,8 @@ openthread_core_files = [
   "mac/mac_filter.hpp",
   "mac/mac_frame.cpp",
   "mac/mac_frame.hpp",
+  "mac/mac_header_ie.cpp",
+  "mac/mac_header_ie.hpp",
   "mac/mac_links.cpp",
   "mac/mac_links.hpp",
   "mac/mac_types.cpp",
@@ -772,6 +774,7 @@ openthread_radio_sources = [
   "instance/instance.cpp",
   "mac/link_raw.cpp",
   "mac/mac_frame.cpp",
+  "mac/mac_header_ie.cpp",
   "mac/mac_types.cpp",
   "mac/sub_mac.cpp",
   "mac/sub_mac_callbacks.cpp",
